Stop using illegal alias name for eval-reader
'= is already used by Clojure's reader and it is not possible to replace it with
our own implementation. I've decided to simply use "eval" as a default alias for
the eval-reader. The official implementation works with "pure"
expressions like "#=(+ 2 3 4)", but doesn't work with Baum's reducers/readers.
For more details of the edge case, see the test case I'll add in this commit.